Allow using the package along with other middlewares
The previous version did not work with other middleware, as it would
inject the shared_results dict in the next middleware instead of the
actual resolver. Moreover, the middleware are instantiated once in
GraphQLView, meaning the shared results lifespan would extend the
request time.

The new version injects the dict in the resolving context instead of
using a global object.